+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Implement `%generated.whatever` generic transform.
+
 ## 0.1.1
 
 - Fix `%edgeql` for modules.